perf(inventory-system): improve current inventory system
The current one is causing to output strings continously, which potentially going to crash the game if the project getting more and more complex. Because of so, I have done some improvement to it:

- Rename global script to uppercase first letter. This is to point out that this is one of the very core script to be used accross other files.
- Rename globally-called function name to reflect the use of it more accurately.
- Improve the conditions handling to how player can click on defined keys to get the correct gun outputs. Currently itis just a signal call without the actual model of each guns.
- Remove the 'remove_item()' function to simplify the scope of this system for now first. After we got the model, we can start working on this.
- Add comments to explain each code lines more effectively to those who also want to work on this.
- Create a custom signal call in 'Player' script for future implement on each guns.
- Remove previous codes that let player to switch guns based on the key inputs they press to this new improved codes.

This is a tiny improvement to the current built for inventory system. Mostly it is for fixing the continously string ouput that potentially can crash the game if the project getting more and more complex.