Fixed bug for player cannot see themself in multi server
The reason that cause this bug is because we did not set the 'Auto Spawn List' value for the new model. That's why even though you can see the new model perfectly in local server, you cannot see the same thing (or even just viewing around) in multi server because we have not set the permission for this new model to be able to load instantly everytime is on a multi server.